requirements.txt
Django==2.2
djangorestframework==3.12.2
Pillow==8.0.1
pytube==9.7.2
pytz==2020.4
sqlparse==0.4.1
typing-extensions==3.7.4.3
